Make protobuf dependancy private
