+++ CMakeLists.txt
@@ -399,6 +399,7 @@ set(TARGET_CORE_SOURCES
     src/utility/LogCollection.cpp
     src/utility/MemoryWrappers.cpp
     src/utility/Serialization.cpp
+    src/utility/ProtoSerialize.cpp
     src/xlink/XLinkConnection.cpp
     src/xlink/XLinkStream.cpp
     src/openvino/OpenVINO.cpp
@@ -731,7 +732,7 @@ if(DEPTHAI_ENABLE_PROTOBUF)
     message(STATUS "Protobuf support enabled")
     add_subdirectory(include/depthai/utility/protos)
     list(APPEND targets_to_export messages)
-    target_link_libraries(${TARGET_CORE_NAME} PUBLIC protobuf::libprotobuf messages)
+    target_link_libraries(${TARGET_CORE_NAME} PRIVATE protobuf::libprotobuf messages)
 endif()
 
 if(DEPTHAI_ENABLE_REMOTE_CONNECTION)
