+++ include/depthai/pipeline/datatype/SpatialImgDetections.hpp
@@ -10,7 +10,6 @@
 #include "depthai/pipeline/datatype/Buffer.hpp"
 #include "depthai/pipeline/datatype/ImgDetections.hpp"
 #include "depthai/pipeline/datatype/SpatialLocationCalculatorConfig.hpp"
-#include "depthai/schemas/SpatialImgDetections.pb.h"
 #include "depthai/utility/ProtoSerializable.hpp"
 #include "depthai/utility/Serialization.hpp"
 
@@ -49,62 +48,19 @@ class SpatialImgDetections : public Buffer, public utility::ProtoSerializable {
         datatype = DatatypeEnum::SpatialImgDetections;
     };
 
-    std::unique_ptr<google::protobuf::Message> getProtoMessage() const override {
-        // create and populate SpatialImgDetections protobuf message
-        auto spatialImgDetections = std::make_unique<proto::spatial_img_detections::SpatialImgDetections>();
-        spatialImgDetections->set_sequencenum(this->sequenceNum);
+    /**
+     * Serialize message to proto buffer
+     *
+     * @returns serialized message
+     */
+    std::vector<std::uint8_t> serializeProto() const override;
 
-        proto::common::Timestamp* ts = spatialImgDetections->mutable_ts();
-        ts->set_sec(this->ts.sec);
-        ts->set_nsec(this->ts.nsec);
-
-        proto::common::Timestamp* tsDevice = spatialImgDetections->mutable_tsdevice();
-        tsDevice->set_sec(this->tsDevice.sec);
-        tsDevice->set_nsec(this->tsDevice.nsec);
-
-        for(const auto& detection : this->detections) {
-            proto::spatial_img_detections::SpatialImgDetection* spatialImgDetection = spatialImgDetections->add_detections();
-
-            // populate SpatialImgDetection.ImgDetection from struct inheritance
-            proto::img_detections::ImgDetection* imgDetection = spatialImgDetection->mutable_detection();
-            imgDetection->set_label(detection.label);
-            imgDetection->set_confidence(detection.confidence);
-            imgDetection->set_xmin(detection.xmin);
-            imgDetection->set_ymin(detection.ymin);
-            imgDetection->set_xmax(detection.xmax);
-            imgDetection->set_ymax(detection.ymax);
-
-            // populate SpatialImgDetection.Point3f
-            proto::spatial_img_detections::Point3f* spatialCoordinates = spatialImgDetection->mutable_spatialcoordinates();
-            spatialCoordinates->set_x(detection.spatialCoordinates.x);
-            spatialCoordinates->set_y(detection.spatialCoordinates.y);
-            spatialCoordinates->set_z(detection.spatialCoordinates.z);
-
-            // populate SpatialImgDetection.SpatialLocationCalculatorConfigData
-            proto::spatial_img_detections::SpatialLocationCalculatorConfigData* boundingBoxMapping = spatialImgDetection->mutable_boundingboxmapping();
-
-            // populate SpatialImgDetection.SpatialLocationCalculatorConfigData.Rect
-            proto::spatial_img_detections::Rect* roi = boundingBoxMapping->mutable_roi();
-            roi->set_x(detection.boundingBoxMapping.roi.x);
-            roi->set_y(detection.boundingBoxMapping.roi.y);
-            roi->set_width(detection.boundingBoxMapping.roi.width);
-            roi->set_height(detection.boundingBoxMapping.roi.height);
-
-            // populate SpatialImgDetection.SpatialLocationCalculatorConfigData.SpatialLocationCalculatorConfigThresholds
-            proto::spatial_img_detections::SpatialLocationCalculatorConfigThresholds* depthTresholds = boundingBoxMapping->mutable_depththresholds();
-            depthTresholds->set_lowerthreshold(detection.boundingBoxMapping.depthThresholds.lowerThreshold);
-            depthTresholds->set_upperthreshold(detection.boundingBoxMapping.depthThresholds.upperThreshold);
-
-            // populate SpatialImgDetection.SpatialLocationCalculatorConfigData.SpatialLocationCalculatorAlgorithm
-            boundingBoxMapping->set_calculationalgorithm(
-                static_cast<proto::spatial_img_detections::SpatialLocationCalculatorAlgorithm>(detection.boundingBoxMapping.calculationAlgorithm));
-
-            // populate SpatialImgDetection.SpatialLocationCalculatorConfigData.stepSize
-            boundingBoxMapping->set_stepsize(detection.boundingBoxMapping.stepSize);
-        }
-
-        return spatialImgDetections;
-    }
+    /**
+     * Serialize schema to proto buffer
+     *
+     * @returns serialized schema
+     */
+    utility::ProtoSerializable::SchemaPair serializeSchema() const override;
 
     DEPTHAI_SERIALIZE(SpatialImgDetections, Buffer::sequenceNum, Buffer::ts, Buffer::tsDevice, detections, transformation);
 };
