+++ src/pipeline/datatype/EncodedFrame.cpp
@@ -1,4 +1,6 @@
 #include "depthai/pipeline/datatype/EncodedFrame.hpp"
+#include "depthai/schemas/EncodedFrame.pb.h"
+#include "../../utility/ProtoSerialize.hpp"
 
 #include "utility/H26xParsers.hpp"
 
@@ -122,45 +124,55 @@ EncodedFrame& EncodedFrame::setProfile(Profile profile) {
     return *this;
 }
 
-std::unique_ptr<google::protobuf::Message> EncodedFrame::getProtoMessage() const {
+
+std::unique_ptr<google::protobuf::Message> getProtoMessage(const EncodedFrame* frame) {
     // Create a unique pointer to the protobuf EncodedFrame message
     auto encodedFrame = std::make_unique<proto::encoded_frame::EncodedFrame>();
 
     // Populate the protobuf message fields with the EncodedFrame data
-    encodedFrame->set_instancenum(this->instanceNum);  // instanceNum -> instancenum
-    encodedFrame->set_width(this->width);
-    encodedFrame->set_height(this->height);
-    encodedFrame->set_quality(this->quality);
-    encodedFrame->set_bitrate(this->bitrate);
-    encodedFrame->set_profile(static_cast<proto::encoded_frame::Profile>(this->profile));  // Profile enum
-    encodedFrame->set_lossless(this->lossless);
-    encodedFrame->set_type(static_cast<proto::encoded_frame::FrameType>(this->type));  // FrameType enum
-    encodedFrame->set_frameoffset(this->frameOffset);                                  // frameOffset -> frameoffset
-    encodedFrame->set_framesize(this->frameSize);                                      // frameSize -> framesize
-    encodedFrame->set_sequencenum(this->sequenceNum);                                  // sequenceNum -> sequencenum
+    encodedFrame->set_instancenum(frame->instanceNum);  // instanceNum -> instancenum
+    encodedFrame->set_width(frame->width);
+    encodedFrame->set_height(frame->height);
+    encodedFrame->set_quality(frame->quality);
+    encodedFrame->set_bitrate(frame->bitrate);
+    encodedFrame->set_profile(static_cast<proto::encoded_frame::Profile>(frame->profile));  // Profile enum
+    encodedFrame->set_lossless(frame->lossless);
+    encodedFrame->set_type(static_cast<proto::encoded_frame::FrameType>(frame->type));  // FrameType enum
+    encodedFrame->set_frameoffset(frame->frameOffset);                                  // frameOffset -> frameoffset
+    encodedFrame->set_framesize(frame->frameSize);                                      // frameSize -> framesize
+    encodedFrame->set_sequencenum(frame->sequenceNum);                                  // sequenceNum -> sequencenum
 
     // Set timestamps
     proto::common::Timestamp* ts = encodedFrame->mutable_ts();
-    ts->set_sec(this->ts.sec);
-    ts->set_nsec(this->ts.nsec);
+    ts->set_sec(frame->ts.sec);
+    ts->set_nsec(frame->ts.nsec);
 
     proto::common::Timestamp* tsDevice = encodedFrame->mutable_tsdevice();
-    tsDevice->set_sec(this->tsDevice.sec);
-    tsDevice->set_nsec(this->tsDevice.nsec);
+    tsDevice->set_sec(frame->tsDevice.sec);
+    tsDevice->set_nsec(frame->tsDevice.nsec);
 
     // Set camera settings
     proto::common::CameraSettings* cam = encodedFrame->mutable_cam();
-    cam->set_exposuretimeus(this->cam.exposureTimeUs);    // exposureTimeUs -> exposuretimeus
-    cam->set_sensitivityiso(this->cam.sensitivityIso);    // sensitivityIso -> sensitivityiso
-    cam->set_lensposition(this->cam.lensPosition);        // lensPosition -> lensposition
-    cam->set_wbcolortemp(this->cam.wbColorTemp);          // wbColorTemp -> wbcolortemp
-    cam->set_lenspositionraw(this->cam.lensPositionRaw);  // lensPositionRaw -> lenspositionraw
+    cam->set_exposuretimeus(frame->cam.exposureTimeUs);    // exposureTimeUs -> exposuretimeus
+    cam->set_sensitivityiso(frame->cam.sensitivityIso);    // sensitivityIso -> sensitivityiso
+    cam->set_lensposition(frame->cam.lensPosition);        // lensPosition -> lensposition
+    cam->set_wbcolortemp(frame->cam.wbColorTemp);          // wbColorTemp -> wbcolortemp
+    cam->set_lenspositionraw(frame->cam.lensPositionRaw);  // lensPositionRaw -> lenspositionraw
 
     // Set the encoded frame data
-    encodedFrame->set_data(this->data->getData().data(), this->data->getData().size());
+    encodedFrame->set_data(frame->data->getData().data(), frame->data->getData().size());
 
     // Return the populated protobuf message
     return encodedFrame;
+}
+
+utility::ProtoSerializable::SchemaPair EncodedFrame::serializeSchema() const {
+    return utility::serializeSchema(getProtoMessage(this));
+}
+
+
+std::vector<std::uint8_t> EncodedFrame::serializeProto() const {
+    return utility::serializeProto(getProtoMessage(this));
 }
 
 }  // namespace dai
